refactor: count 계열 JPA 메서드의 반환 타입 변경
- PK 타입과 동일한 `Long`으로 변경

+++ src/main/kotlin/com/barostartbe/domain/user/repository/AccessLogRepository.kt
@@ -21,5 +21,5 @@ interface AccessLogRepository : JpaRepository<AccessLog, Long> {
             GROUP BY DATE_SUB(log_date, INTERVAL rnk DAY)
         ) streaks
     """, nativeQuery = true)
-    fun findMaxConsecutiveDays(@Param("userId") userId: Long): Int
+    fun findMaxConsecutiveDays(@Param("userId") userId: Long): Long
 }
